fix: fix warning in loop (#1)
* fix: fix warning in loop

Description
------------

Root cause
------------

Fix solution
------------

How Has This Been Tested?
------------

Release Note
------------

```release-note
```

* feat: add 206 reverse linked list with iterative and recursive solutions

- Add iterative solution using three pointers
- Add recursive solution with backtracking
- Include comprehensive test cases for both approaches
- Fix missing <algorithm> include in test_utils.h



* ci: add GitHub Actions workflow for PR testing

- Run CMake build and ctest on push to main and pull requests
- Ensure all tests pass before merging



* ci: add comprehensive C++ checks

- clang-format: code style consistency check
- clang-tidy: static analysis for potential bugs
- Multi-compiler: test with both GCC and Clang
- AddressSanitizer: runtime memory error detection



* style: enhance clang-format with Facebook-style conventions

- C++17 standard
- 4-space indentation with proper access modifier offset
- Attach braces style
- Left pointer alignment
- Include sorting with priority groups
- Constructor initializer formatting



* docs: add clang-format commands to AGENTS.md



---------

+++ 0001_two_sum/solution.cpp
@@ -8,7 +8,7 @@ public:
     vector<int> twoSum(vector<int>& nums, int target) {
         unordered_map<int, int> numMap;
         
-        for (int i = 0; i < nums.size(); ++i) {
+        for (int i = 0; i < static_cast<int>(nums.size()); ++i) {
             int complement = target - nums[i];
             
             if (numMap.find(complement) != numMap.end()) {
